move extension components
