+++ src/packages/core/components/index.ts
@@ -6,8 +6,6 @@ export * from './body-layout/body-layout.element.js';
 export * from './code-block/index.js';
 export * from './dropdown/index.js';
 export * from './entity-actions-bundle/index.js';
-export * from './extension-slot/index.js';
-export * from './extension-with-api-slot/index.js';
 export * from './footer-layout/index.js';
 export * from './header-app/index.js';
 export * from './history/index.js';
