+++ src/packages/core/extension-registry/components/extension-slot/extension-slot.element.ts
@@ -1,4 +1,4 @@
-import { umbExtensionsRegistry } from '../../extension-registry/index.js';
+import { umbExtensionsRegistry } from '../../registry.js';
 import { UmbLitElement } from '@umbraco-cms/backoffice/lit-element';
 import type { TemplateResult } from '@umbraco-cms/backoffice/external/lit';
 import { css, repeat, customElement, property, state, html } from '@umbraco-cms/backoffice/external/lit';
