+++ src/packages/core/extension-registry/components/extension-slot/extension-slot.test.ts
