+++ src/packages/core/extension-registry/components/extension-slot/index.ts
