+++ src/packages/core/extension-registry/components/extension-with-api-slot/extension-with-api-slot.test.ts
