+++ src/packages/core/extension-registry/components/extension-with-api-slot/index.ts
