+++ src/packages/core/extension-registry/components/index.ts
@@ -1,0 +1,2 @@
+export * from './extension-slot/index.js';
+export * from './extension-with-api-slot/index.js';
